refactor: move useDelayable to src
